+++ src/backend/index.js
@@ -140,6 +140,15 @@ function serializer(replacer, cycleReplacer) {
     }
 }
 
+app.get('/testdata', function (req, res) {
+    res.send("TEST DATA FROM BACKEND");
+});
+
+app.post('/testdata', function (req, res) {
+    console.log(">>> TEST DATA RECEIVED: ", req.body);
+    res.send("OK");
+});
+
 app.get('/', function (req, res) {
     res.sendFile(vuePath + "index.html");
 });