+++ src/pages/PageStructure.js
@@ -56,7 +56,7 @@ export default {
                     component: AddRoutePage,
                     subpages: []
                 }, {
-                    path: "showroute",
+                    path: "showrouteoffering",
                     name: "Show Route",
                     component: AddRoutePage,
                     subpages: []
@@ -96,7 +96,7 @@ export default {
                     component: AddRoutePage,
                     subpages: []
                 }, {
-                    path: "showroute",
+                    path: "showrouteconsumption",
                     name: "Show Route",
                     component: AddRoutePage,
                     subpages: []