+++ CHANGELOG.md
@@ -2,14 +2,17 @@
 All notable changes to this project will be documented in this file.
 (Skipped major version 1, 2 and 3 to match versioning of IDS DataSpaceConnector)
 
-## [X.X.X]
+## [9.1.0] - 2022-03-25 (compatible with DSC 7.0.0)
 
 ### Added
-
-### Changes
+- Data Offering: Add local data resources
+- Backend Connections: Source type "Other"
+- Added route creator (Data Offering & Consumption) with apps
+- Data Consumption: Dispatch data via routes
 
 ### Fixes
 - Raised comparison of variable from value to type level
+- Removed sensitive data in HTTP response
 
 ## [9.0.0] - 2022-02-07 (compatible with DSC 7.0.0)